requirements.txt
beautifulsoup4==4.6.3
Django==2.1
django-livereload-server==0.3.2
django-webpack-loader==0.6.0
djangorestframework==3.8.2
pytz==2018.5
six==1.11.0
tornado==5.1
